Refacotr

+++ src/term.rs
@@ -17,7 +17,7 @@ struct Writer {
     pub stdout: RawTerminal<Stdout>, 
     pub chars: Vec<char>,
     pub widths: Vec<u8>,
-    pub ch_ptr: i32,
+    pub ch_ptr: usize,
     pub hist_ptr: usize,
 }
 
@@ -65,11 +65,13 @@ impl Writer {
             return;
         }
 
+        /* THERE MAY BE A BUG */
         let h = &history[self.hist_ptr as usize];
         self.rewrite_line(left, y, h.commandline.to_string());
         self.chars = h.commandline.chars().collect();
         self.widths = h.charwidths.clone();
-        self.ch_ptr = self.widths.len() as i32;
+        self.ch_ptr = self.widths.len();
+//        println!("{:?}\n{:?}\n{:?}", self.chars, self.widths, self.ch_ptr);
     }
 }
 
@@ -81,7 +83,7 @@ pub fn prompt(text: &String) -> u16 {
     prompt.len().try_into().unwrap()
 }
 
-fn left_ch_ptr(pos: i32) -> i32 {
+fn left_ch_ptr(pos: usize) -> usize {
     if pos == 0 {
         0
     }else{
@@ -107,18 +109,18 @@ pub fn read_line(left: u16, history: &mut Vec<History>) -> String{
             },
             event::Key::Left => {
                 writer.ch_ptr = left_ch_ptr(writer.ch_ptr);
-                if x-writer.widths[writer.ch_ptr as usize] as u16 > left {
-                    writer.cur_move(x-writer.widths[writer.ch_ptr as usize] as u16, y);
+                if x-writer.widths[writer.ch_ptr] as u16 > left {
+                    writer.cur_move(x-writer.widths[writer.ch_ptr] as u16, y);
                 };
             },
             event::Key::Right => {
-                if writer.chars.len() as i32 > writer.ch_ptr+1 {
+                if writer.chars.len() > writer.ch_ptr+1 {
                     writer.ch_ptr += 1;
-                    writer.cur_move(x+writer.widths[writer.ch_ptr as usize] as u16, y);
+                    writer.cur_move(x+writer.widths[writer.ch_ptr] as u16, y);
                 }else{
                     let line_len: u16 = writer.widths.iter().fold(0, |line_len, w| line_len + (*w as u16));
                     writer.cur_move(left+line_len+1, y);
-                    writer.ch_ptr = writer.chars.len() as i32;
+                    writer.ch_ptr = writer.chars.len();
                 };
             },
             event::Key::Backspace => {
@@ -127,11 +129,11 @@ pub fn read_line(left: u16, history: &mut Vec<History>) -> String{
                 };
 
                 writer.ch_ptr = left_ch_ptr(writer.ch_ptr);
-                writer.chars.remove(writer.ch_ptr as usize);
+                writer.chars.remove(writer.ch_ptr);
                 writer.rewrite_line(left, y, writer.chars.iter().collect());
 
-                if x - writer.widths[writer.ch_ptr as usize] as u16 >= left {
-                    writer.cur_move(x-writer.widths[writer.ch_ptr as usize] as u16, y);
+                if x - writer.widths[writer.ch_ptr] as u16 >= left {
+                    writer.cur_move(x-writer.widths[writer.ch_ptr] as u16, y);
                 }
             },
             event::Key::Char(c) => {
@@ -140,17 +142,17 @@ pub fn read_line(left: u16, history: &mut Vec<History>) -> String{
                         writer.chars.push(c);
                         break;
                     }
-                    writer.chars.insert(writer.ch_ptr as usize, c);
+                    writer.chars.insert(writer.ch_ptr, c);
                     writer.ch_ptr += 1;
 
                     /* output the line before the cursor */
-                    writer.rewrite_line(left, y, writer.chars[0..writer.ch_ptr as usize].iter().collect());
+                    writer.rewrite_line(left, y, writer.chars[0..writer.ch_ptr].iter().collect());
                     let (new_x, new_y) = writer.cursor_pos();
-                    writer.widths.insert(writer.ch_ptr as usize -1, (new_x - x) as u8);
+                    writer.widths.insert(writer.ch_ptr-1, (new_x - x) as u8);
 
                     /* output the line after the cursor */
                     write!(writer.stdout, "{}{}",
-                           writer.chars[writer.ch_ptr as usize ..].iter().collect::<String>(), 
+                           writer.chars[writer.ch_ptr..].iter().collect::<String>(), 
                            termion::cursor::Goto(new_x, new_y),
                     ).unwrap();
                     writer.stdout.flush().unwrap();
